fix: create amp track hook
